lib/python/requirements_dev.txt
nose
nose-exclude==0.4.1
rednose
mock==1.0.1
coverage
six==1.10.0

nats-client==0.2.4

asyncio-nats-client==0.3.1
aiohttp==0.22.3

-r requirements.txt
